Remove broken icon rendering for loading="lazy" images while loading
Non-lazy images don't show this icon. It's incorrect to show such
an icon just because an image is loading. In addition, it messes
up the intrinsic sizing heuristics for aspect ratio.

Both lazy and non-lazy images still show a broken image icon when
images fail to load via HTTP failure.

Bug: 997606, 1045745

Change-Id: I5dc6bac50a1c43395057e37547864e9e3fc13b6d