[LayoutNG] Fix text-overflow: ellipsis with scrolling
When we have "text-overflow: ellipsis" on an element we only truncate
the text once (during the layout) and not after a scrolling happens.
So if we perform any kind of scrolling after the initial layout,
the position of the ellipsis doesn't change and we cannot see
the rest of the text.

This patch fixes the issue by forcing a relayout during scrolling
in PaintLayerScrollableArea::UpdateScrollOffset() if the element
has to truncate the overflow.

The patch also modifies NGInlineLayoutAlgorithm::CreateLine()
to take into account the scrolling offset and pass this information to
to NGLineTruncator::TruncateLine().

Apart from the new tests added on this patch, the following one
starts to pass on Linux and Windows: css/css-ui/text-overflow-021.html
(not passing on Mac due to rounding issues related to fonts).

Bug=745905
TEST=external/wpt/css/css-overflow/text-overflow-scroll-001.html
TEST=external/wpt/css/css-overflow/text-overflow-scroll-rtl-001.html
TEST=external/wpt/css/css-overflow/text-overflow-scroll-vertical-lr-001.html
TEST=external/wpt/css/css-overflow/text-overflow-scroll-vertical-lr-rtl-001.html
TEST=external/wpt/css/css-overflow/text-overflow-scroll-vertical-rl-001.html
TEST=external/wpt/css/css-overflow/text-overflow-scroll-vertical-rl-rtl-001.html

Change-Id: I975ca86dbdfedf92875c08736539fb2f33f68c54
Bug: 745905